vlc_fixup: pollfd.(r)events is short int
As per POSIX (and Winsock).

(cherry picked from commit 6c172726cb0146864edbb7703e3eb2979afea87c)

+++ include/vlc_fixups.h
@@ -406,8 +406,8 @@ enum
 struct pollfd
 {
     int fd;
-    unsigned events;
-    unsigned revents;
+    short events;
+    short revents;
 };
 #endif
 #ifndef HAVE_POLL
